fix(submittoken): inverted meaning of submittoken, undefined (falsy) now indicates a valid form
Inversion was done to provide a changing submittoken when users press the submit button repeatedly
even when the form has errors. This allows for side-effect (useEffect etc) to look to the
submittoken and do additional measure like indicating the errors to the users

BREAKING CHANGE: submittoken was change from a boolean type, to a string? type, and the
falsy/truthy-values was inverted.