feat(deps): bump immer-dependency to 9.x.x
Should in theory cause any problem, but out of an abundance of caution we mark this as breaking.

BREAKING CHANGE: major bump of immer, 3.x.x to 9.x.x. Shouldn't in theory be breaking, but since consumers may already use immer we take extra caution here and bump major anyway